+++ CHANGES.rst
@@ -1,10 +1,18 @@
+0.5.0 18/06/2018
+================
+* add support for django 2.0
+* add `query_params` property to allow handling POST request
+
+
 0.4.0 29/05/2017
 ================
 * add '__inarray' and  '__int_inarray' lookup to handle json/arrays lookup both str/int
 
+
 0.3.0 10/10/16
 ==============
 * add '_distinct' parameter to enable '.distinct()' queries
+
 
 0.2.0 19/09/16
 ==============